+++ portal-impl/src/com/liferay/portlet/documentlibrary/util/XugglerAutoInstallHelper.java
@@ -14,16 +14,16 @@
 
 package com.liferay.portlet.documentlibrary.util;
 
+import com.liferay.petra.process.ProcessException;
+import com.liferay.petra.reflect.ReflectionUtil;
+import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.configuration.Filter;
 import com.liferay.portal.kernel.io.DummyOutputStream;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.process.ClassPathUtil;
-import com.liferay.portal.kernel.process.ProcessException;
 import com.liferay.portal.kernel.util.OSDetector;
 import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.ReflectionUtil;
 import com.liferay.portal.kernel.util.SetUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.xuggler.Xuggler;
@@ -41,7 +41,9 @@ import java.lang.reflect.Field;
 import java.net.URL;
 import java.net.URLClassLoader;
 
+import java.util.Collections;
 import java.util.Iterator;
+import java.util.LinkedHashSet;
 import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.Callable;
@@ -68,9 +70,10 @@ public class XugglerAutoInstallHelper {
 
 		if (xugglerJarFile == null) {
 			_log.error(
-				"Xuggler auto install is not supported on system: " +
-					System.getProperty("os.name") + "/" +
-						System.getProperty("os.arch"));
+				StringBundler.concat(
+					"Xuggler auto install is not supported on system: ",
+					System.getProperty("os.name"), "/",
+					System.getProperty("os.arch")));
 
 			return;
 		}
@@ -197,7 +200,19 @@ public class XugglerAutoInstallHelper {
 
 		ClassLoader contextClassLoader = currentThread.getContextClassLoader();
 
-		Set<URL> urls = ClassPathUtil.getClassPathURLs(contextClassLoader);
+		Set<URL> urls = new LinkedHashSet<>();
+
+		ClassLoader classLoader = contextClassLoader;
+
+		while (classLoader != null) {
+			if (classLoader instanceof URLClassLoader) {
+				URLClassLoader urlClassLoader = (URLClassLoader)classLoader;
+
+				Collections.addAll(urls, urlClassLoader.getURLs());
+			}
+
+			classLoader = classLoader.getParent();
+		}
 
 		Iterator<URL> iterator = urls.iterator();
 
@@ -211,18 +226,18 @@ public class XugglerAutoInstallHelper {
 
 				Matcher matcher = _pattern.matcher(file.getName());
 
-				if (matcher.matches()) {
-					if (jarFiles.contains(matcher.replaceAll("$1$2"))) {
-						file.delete();
+				if (matcher.matches() &&
+					jarFiles.contains(matcher.replaceAll("$1$2"))) {
 
-						iterator.remove();
-					}
+					file.delete();
+
+					iterator.remove();
 				}
 			}
 		}
 
 		URLClassLoader urlClassLoader = new URLClassLoader(
-			urls.toArray(new URL[urls.size()]), null);
+			urls.toArray(new URL[0]), null);
 
 		currentThread.setContextClassLoader(urlClassLoader);
